refactor qrgenerator
