--- src/components/QRGenerator/index.tsx
@@ -1,3 +1,0 @@
-import { QrCodeGenerator } from "./QrGenerator";
-
-export { QrCodeGenerator };
